aaaaaaaaa complex ya ebal pochemu vse tak slohzno

+++ include/time.h
@@ -2,10 +2,10 @@
 
 #include <chrono>
 
-double time() {
+long double time() {
     static chrono::time_point<chrono::high_resolution_clock> start_clock_time = chrono::high_resolution_clock::now();
     static chrono::time_point<chrono::high_resolution_clock> now_time;
     now_time = chrono::high_resolution_clock::now();
-    chrono::duration<double> delta = now_time - start_clock_time;
+    chrono::duration<long double> delta = now_time - start_clock_time;
     return delta.count();
 }
